+++ src/main/java/com/sparta/eduwithme/domain/comment/dto/CommentRoomDto.java
@@ -14,5 +14,5 @@ public class CommentRoomDto {
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
     private String roomName;
-    private Integer questionOrderInRoom;
+    private Long questionOrderInRoom;
 }
