+++ src/main/java/com/sparta/eduwithme/domain/question/dto/QuestionResponseDto.java
@@ -14,7 +14,7 @@ public class QuestionResponseDto {
     private final String category;
     private final String difficulty;
     private final Long point;
-    private final Integer orderInRoom;
+    private final Long orderInRoom;
     private final LocalDateTime createdAt;
     private final LocalDateTime updatedAt;
